pyproject.toml
[project]
name = "h1blinkedincompanynamecrewai"
version = "0.1.0"
description = "h1blinkedincompanynamecrewai using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.2.1",
    "requests>=2.31.0,<3.0.0",
    "slack-sdk>=3.23.0,<4.0.0",
    "python-dotenv>=1.0.0,<2.0.0"
]

[project.scripts]
h1blinkedincompanynamecrewai = "h1blinkedincompanynamecrewai.main:run"
run_crew = "h1blinkedincompanynamecrewai.main:run"
train = "h1blinkedincompanynamecrewai.main:train"
replay = "h1blinkedincompanynamecrewai.main:replay"
test = "h1blinkedincompanynamecrewai.main:test"
run_with_trigger = "h1blinkedincompanynamecrewai.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
